Share SOTA comparison JSON serialization
